chore(client): move files to different folders, change src path
