+++ client/caption/delete-caption.js
