--- client/index.css
@@ -1,348 +1,0 @@
-/* J.M */
-html, body {
-    width: 100%;
-    height:100%;
-}
-
-
-/* Background Image J.M */ 
-.bkgnd-container{
-    background: url(/client/img/purpgalaxy.jpg);
-    color: rgb(0, 0, 0);
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    position: relative;
-    /* G.A Start */
-    /* overflow: hidden;
-    z-index: -1;
-    /* G.A End */
-
-}
-
-/* Background Color overlay J.M START */
-.color-overlay{
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(-45deg, #eedc52, #cc2968, #003571, rgb(0, 0, 1));
-    background-size: 400% 400%;
-    animation: gradient 20s ease infinite;
-    opacity: .2;
-    position: relative;
-}
-
-
-    @keyframes gradient {
-        0% {
-            background-position: 0% 50%;
-        }
-        50% {
-            background-position: 100% 50%;
-        }
-        100% {
-            background-position: 0% 50%;
-        }
-    }
-/* Background color overlay J.M END */
-
-
-/* OTTER SPACE Image J.M START */
-.flier{
-    pointer-events: none;
-    opacity: .7;
-}
-
-.flier > * {
-    /* Adjust animation duration to change the element’s speed J.M */
-        animation: fly 26s linear infinite;
-        pointer-events: none !important;
-        top: 0;
-        left: 0;
-        transform: translateX(-120%) translateY(-120%) rotateZ(0);
-        position: fixed;
-        animation-delay: 1s;
-        z-index: 999999;
-    }
-    
-     /* Keyframe values control where the element will begin
-        and end its trajectory across the screen. Each rule
-        represents a path the element follows across the screen. */
-    
-/* MEME IMAGE G.A Start */
-.meme-image {
-    position: absolute;
-    top: 35%;
-    left: 10%;
-}
-
-.meme-image canvas {
-    background:rgba(255, 255, 255, 0.2);
-    /* height: 400px;
-    width: 800px; */
-}
-
-.upload-image-div {
-    color: white;
-}
-/* MEME IMAGE G.A End */
-    
-
-/* Generator Buttons J.M START */
-#btn{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto', sans-serif;
-    opacity: 1;
-    border: none;
-    border-radius: 6px;
-    height: 40px;
-    width: 180px;
-}
-
-#btn:hover {
-    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-    color: rgb(255, 255, 255);
-    background-color: #d353f3f6;
-    -webkit-transform: translateY(-5px);
-	transform: translateY(-5px);
-}
-
-
-#uploadMeme{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto', sans-serif;
-    opacity: 1;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-
-#uploadMeme:hover{
-    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-    color: rgb(255, 255, 255);
-    background-color: #d353f3f6;
-    -webkit-transform: translateY(-5px);
-	transform: translateY(-5px);
-}
-
-#topText{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto', sans-serif;
-    font-size: 16px;
-    opacity: .7;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-
-#botText{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto' sans-serif;
-    font-size: 16px;
-    opacity: .7;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-
-.randombtn{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto' sans-serif;
-    font-size: 16px;
-    opacity: 1;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-.randombtn:hover{
-        box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-        color: rgb(255, 255, 255);
-        background-color: #d353f3f6;
-        -webkit-transform: translateY(-5px);
-        transform: translateY(-5px);
-}
-
-.savebtn{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto' sans-serif;
-    font-size: 16px;
-    opacity: 1;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-.savebtn:hover{
-    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-    color: rgb(255, 255, 255);
-    background-color: #d353f3f6;
-    -webkit-transform: translateY(-5px);
-	transform: translateY(-5px);
-}
-
-
-
-  
-
-#Canvas{
-    width: 250px;
-    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-  text-align: center;
-}
-
-
-/* Generator Buttons J.M END */
-
-
-
-
-
-
-    @keyframes fly {
-    
-        98.001%, 0% {
-                    display: block;
-            transform: translateX(-200%) translateY(100vh) rotateZ(0deg)
-        }
-    
-        15% {
-            transform: translateX(100vw) translateY(-100%) rotateZ(180deg)
-        }
-    
-        15.001%, 18% {
-            transform: translateX(100vw) translateY(-30%) rotateZ(0deg)
-        }
-    
-        40% {
-            transform: translateX(-200%) translateY(3vh) rotateZ(-180deg)
-        }
-    
-        40.001%, 43% {
-            transform: translateX(-200%) translateY(-100%) rotateZ(-180deg)
-        }
-    
-        65% {
-            transform: translateX(100vw) translateY(50vh) rotateZ(0deg)
-        }
-    
-        65.001%, 68% {
-            transform: translateX(20vw) translateY(-200%) rotateZ(180deg)
-        }
-    
-        95% {
-            transform: translateX(10vw) translateY(100vh) rotateZ(0deg)
-        }
-    }
-/* OTTER SPACE IMAGE J.M END */
-
-
-
-/* Info/How to use Pop-up box J.M START */
-/* The Modal (background) */
-.modal {
-    display: none; /* Hidden by default */
-    position: fixed; /* Stay in place */
-    z-index: 1; /* Sit on top */
-    padding-top: 250px; /* Location of the box */
-    left: 0px;
-    top: 0;
-    width: 100%; /* Full width */
-    height: 100%; /* Full height */
-    overflow: auto; /* Enable scroll if needed */
-    background-color: rgb(0,0,0); /* Fallback color */
-    background-color: rgba(0, 0, 0, 0.688); /* Black w/ opacity */
-  }
-  
-/* Modal Content */
-.modal-content {
-    background-color: #fcf0ff;
-    margin: auto;
-    padding: 20px;
-    border: 1px solid #888;
-    width: 700px;
-    height: 400px;
-    box-shadow:1px 1px 20px rgb(100, 0, 137), 1px 0px 20px rgb(155, 0, 182);
-    overflow-x: scroll;
-  }
-
-  .modal-img{
-    background-image: url(/client/img/ufo\ and\ shaq\ 3.png);
-     background-size: 100%;
-     background-position:right;
-    overflow: scroll; 
-    background-repeat: no-repeat;
-  }
-
-  /* @media screen and (max-width: 670px) {
-    .modal-img {
-        position: static;
-        overflow: auto;
-        width: 100%;
-        height: 100%;
-        border-left: none;
-        clear: both;
-    }
-}
-@media screen and (max-width: 380px) {
-    .modal-img {
-        width: 380px;
-    }
-} */
-
-
-
-  /* The Close Button */
-  .close {
-    color: #00000092;
-    float: left;
-    font-size: 28px;
-    font-weight: bold;
-    text-shadow:1px 1px 10px rgb(197, 38, 255), 1px 1px 20px rgb(111, 0, 152);
-  }
-
-  .close:hover,
-  .close:focus {
-    color: #000;
-    text-decoration: none;
-    cursor: pointer;
-  }
-
-  .info-container {
-      background-color: #d0b5b500;
-      opacity: 1
-      ;
-      position:absolute;
-      bottom: 6%;
-      right:45%;
-  }
-
-#myBtn{
-    background-color: rgb(227, 222, 232);
-    font-family: 'Roboto' sans-serif;
-    font-size: 16px;
-    opacity: 1;
-    border: none;
-    border-radius: 8px;
-    height: 40px;
-    width: 180px;
-}
-
-#myBtn:hover{
-    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
-    color: rgb(255, 255, 255);
-    background-color: #d353f3f6;
-    -webkit-transform: translateY(-5px);
-	transform: translateY(-5px);
-}
-
-.head-text{
-    font-family: 'Chau Philomene One', sans-serif;
-    font-size: 42px;
-    text-shadow:1px 1px 10px rgb(197, 38, 255), 1px 1px 10px rgb(111, 0, 152);
-}
-/* Info/How to use Pop-up box J.M END */
-
