+++ client/meme/meme.css
@@ -126,36 +126,258 @@ html, body {
     }
 /* OTTER SPACE IMAGE J.M END */
 
+/* Generator Buttons J.M START */
+#btn{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto', sans-serif;
+    opacity: 1;
+    border: none;
+    border-radius: 6px;
+    height: 40px;
+    width: 180px;
+}
+
+#btn:hover {
+    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+    color: rgb(255, 255, 255);
+    background-color: #d353f3f6;
+    -webkit-transform: translateY(-5px);
+	transform: translateY(-5px);
+}
+
+
+#uploadMeme{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto', sans-serif;
+    opacity: 1;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
+}
+
+#uploadMeme:hover{
+    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+    color: rgb(255, 255, 255);
+    background-color: #d353f3f6;
+    -webkit-transform: translateY(-5px);
+	transform: translateY(-5px);
+}
+
+#topText{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto', sans-serif;
+    font-size: 16px;
+    opacity: .7;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
+}
+
+#botText{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto' sans-serif;
+    font-size: 16px;
+    opacity: .7;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
+}
+
+.randombtn{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto' sans-serif;
+    font-size: 16px;
+    opacity: 1;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
+}
+.randombtn:hover{
+        box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+        color: rgb(255, 255, 255);
+        background-color: #d353f3f6;
+        -webkit-transform: translateY(-5px);
+        transform: translateY(-5px);
+}
+
+.savebtn{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto' sans-serif;
+    font-size: 16px;
+    opacity: 1;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
+}
+.savebtn:hover{
+    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+    color: rgb(255, 255, 255);
+    background-color: #d353f3f6;
+    -webkit-transform: translateY(-5px);
+	transform: translateY(-5px);
+}
 
 
 
+  
+
+#Canvas{
+    width: 250px;
+    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+  text-align: center;
+}
+
+
+/* Generator Buttons J.M END */
+
+    @keyframes fly {
+    
+        98.001%, 0% {
+                    display: block;
+            transform: translateX(-200%) translateY(100vh) rotateZ(0deg)
+        }
+    
+        15% {
+            transform: translateX(100vw) translateY(-100%) rotateZ(180deg)
+        }
+    
+        15.001%, 18% {
+            transform: translateX(100vw) translateY(-30%) rotateZ(0deg)
+        }
+    
+        40% {
+            transform: translateX(-200%) translateY(3vh) rotateZ(-180deg)
+        }
+    
+        40.001%, 43% {
+            transform: translateX(-200%) translateY(-100%) rotateZ(-180deg)
+        }
+    
+        65% {
+            transform: translateX(100vw) translateY(50vh) rotateZ(0deg)
+        }
+    
+        65.001%, 68% {
+            transform: translateX(20vw) translateY(-200%) rotateZ(180deg)
+        }
+    
+        95% {
+            transform: translateX(10vw) translateY(100vh) rotateZ(0deg)
+        }
+    }
+/* OTTER SPACE IMAGE J.M END */
 
 
 
-
-
-
-
-
-
-/* table G.S.A Start */
-.caption-table{
-    width: 100vh;
-    border: 1px solid black;
-    margin: 30 30 30 30;
+/* Info/How to use Pop-up box J.M START */
+/* The Modal (background) */
+.modal {
+    display: none; /* Hidden by default */
+    position: fixed; /* Stay in place */
+    z-index: 1; /* Sit on top */
+    padding-top: 250px; /* Location of the box */
+    left: 0px;
+    top: 0;
+    width: 100%; /* Full width */
+    height: 100%; /* Full height */
+    overflow: auto; /* Enable scroll if needed */
+    background-color: rgb(0,0,0); /* Fallback color */
+    background-color: rgba(0, 0, 0, 0.688); /* Black w/ opacity */
+  }
+  
+/* Modal Content */
+.modal-content {
+    background-color: #fcf0ff;
+    margin: auto;
     padding: 20px;
-    justify-content: center;
-    align-items: center;
+    border: 1px solid #888;
+    width: 700px;
+    height: 400px;
+    box-shadow:1px 1px 20px rgb(100, 0, 137), 1px 0px 20px rgb(155, 0, 182);
+    overflow-x: scroll;
+  }
+
+  .modal-img{
+    background-image: url(/client/img/ufo\ and\ shaq\ 3.png);
+     background-size: 100%;
+     background-position:right;
+    overflow: scroll; 
+    background-repeat: no-repeat;
+  }
+
+  /* @media screen and (max-width: 670px) {
+    .modal-img {
+        position: static;
+        overflow: auto;
+        width: 100%;
+        height: 100%;
+        border-left: none;
+        clear: both;
+    }
+}
+@media screen and (max-width: 380px) {
+    .modal-img {
+        width: 380px;
+    }
+} */
+
+
+  /* The Close Button */
+  .close {
+    color: #00000092;
+    float: left;
+    font-size: 28px;
+    font-weight: bold;
+    text-shadow:1px 1px 10px rgb(197, 38, 255), 1px 1px 20px rgb(111, 0, 152);
+  }
+
+  .close:hover,
+  .close:focus {
+    color: #000;
+    text-decoration: none;
+    cursor: pointer;
+  }
+
+  .info-container {
+      background-color: #d0b5b500;
+      opacity: 1
+      ;
+      position:absolute;
+      bottom: 6%;
+      right:45%;
+  }
+
+#myBtn{
+    background-color: rgb(227, 222, 232);
+    font-family: 'Roboto' sans-serif;
+    font-size: 16px;
+    opacity: 1;
+    border: none;
+    border-radius: 8px;
+    height: 40px;
+    width: 180px;
 }
 
-tr, th, td{
-    width: 20vh;
-    padding: 10px;
-    justify-content: center;
-    align-items: center;
-    border: 1px solid black;
+#myBtn:hover{
+    box-shadow:1px 1px 10px rgb(100, 0, 137), 1px 1px 10px rgb(155, 0, 182);
+    color: rgb(255, 255, 255);
+    background-color: #d353f3f6;
+    -webkit-transform: translateY(-5px);
+	transform: translateY(-5px);
 }
 
-/* table G.S.A End */
+.head-text{
+    font-family: 'Chau Philomene One', sans-serif;
+    font-size: 42px;
+    text-shadow:1px 1px 10px rgb(197, 38, 255), 1px 1px 10px rgb(111, 0, 152);
+}
+/* Info/How to use Pop-up box J.M END */
+
+
